% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ArrowHead}
\alias{ArrowHead}
\title{The shape of arrow heads.}
\format{An object of class \code{data.frame} with 100 rows and 84 columns.}
\source{
\url{http://timeseriesclassification.com/description.php?Dataset=ArrowHead}
}
\usage{
ArrowHead
}
\description{
A dataset containing the outline of arrow heads as functional covariable.
This is a subset of the "ArrowHead" data set of the UCR TSC repository.
}
\details{
The arrowhead data consists of outlines of the images of arrowheads.
The shapes of the projectile points are converted into a time series using
the angle-based method. The classification of projectile points is an
important topic in anthropology. The classes are based on shape distinctions
such as the presence and location of a notch in the arrow. The problem in the
repository is a length normalised version of that used in Ye09shapelets.
The three classes are called "Avonlea", "Clovis" and "Mix".
}
\section{Format A data frame with 100 rows (=observations) and 84 variables}{

\describe{
  \item{col 1:83}{shape of a projectile as functional observation.}
  \item{target}{encoding the class of the projectile.}
}
}

\keyword{datasets}
