% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdataTransform.R
\name{fdataTransform}
\alias{fdataTransform}
\title{Create a preprocessing pipeline function}
\usage{
fdataTransform(grid, nderiv, derived, evenly.spaced, no.missing, deriv.method,
  ...)
}
\arguments{
\item{grid, nderiv, derived, evenly.spaced, no.missing, deriv.method}{see
\code{\link{classiKnn}}}

\item{...}{additional arguments to fda::smooth.basis}
}
\value{
Pipeline function taking one argument \code{fdata}. The returned
function carries out all the preprocessing needed for the calling model
of class \code{\link{classiKnn}}.
}
\description{
Internal function, documented due to the importance of its
concept. Creates a pipeline function to do all the
preprocessing needed in \code{\link{classiKnn}} and \code{\link{classiKernel}}.
This is helpful to ensure that the data preprocessing (imputation of missing
values, derivation) is carried out in exactly the same way for the
training and the test set in \code{predict.classiKnn} and
\code{predict.classiKernel}.
}
