% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Growth}
\alias{Growth}
\title{Berkeley Growth Study Data (regular grid)}
\format{A list made up of \describe{
\item{ID}{Factor of length 93 containing the subject IDs}
\item{sex}{Factor encoding the sex of children with values in \code{c(`male`, `female`)}}
\item{height}{A 93 x 31 matrix giving the height in cm of 93 children at 31 ages}
}
 @inheritSection Growth_irregular references
 @seealso Growth_irregular}
\description{
A data frame containing the heights of 39 boys and 54 girls from age 1 to 18,
the ages at which they were collected.
}
\details{
Data and description was reformatted from the \code{\link[fda:fda-package]{fda}} package
to be observed on a regular grid in one year steps.
}
