% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ker_choices.R
\name{ker.choices}
\alias{ker.choices}
\title{Name all implemented kernel functions}
\usage{
ker.choices()
}
\description{
\code{ker.choices} is a function returning the names of  all
kernel functions that are currently implemented in the
\code{link{classiFunc}}-package and can be used for the argument
\code{ker} in \code{\link{classiKernel}}.
}
