% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDistMat.R
\name{metric.choices}
\alias{metric.choices}
\title{Name all metrics}
\usage{
metric.choices()
}
\description{
\code{metric.choices} is a function returning all (semi-)metrics that are
currently implemented in the \code{link{classiFunc}}-package.
}
