% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Phoneme}
\alias{Phoneme}
\title{Phonetic Time Series.}
\format{An object of class \code{data.frame} with 100 rows and 65 columns.}
\source{
\url{http://timeseriesclassification.com/description.php?Dataset=Phoneme}
}
\usage{
Phoneme
}
\description{
A dataset containing the audio files of english words.
}
\details{
This data set is a subsample of the data used in the paper Dual-domain
Hierarchical Classification of Phonetic Time Series. Each series is extracted
from the segmented audio collected from Google Translate, oxforddictionaries.com
and the Merrriam-Webster online dictionary. Each of these sources have
different features. Audio files collected from Google translate, Oxford, and
Merrriam-Webster dictionaries are recorded at 22050, 44100 and 11025 samples
per second respectively. All of them have male and female speakers in
different ratios. The Oxford dictionary includes British and American accent
pronunciation for each word. After data collection, they segment waveforms of
the words to generate phonemes using the Forced Aligner tool from the Penn
Phonetics Laboratory.
}
\section{Format A data frame with 100 rows (=observations) and 65 variables}{

\describe{
  \item{col 1:64}{one functional observation.}
  \item{target}{encoding the word of the functional observation}
}
}

\keyword{datasets}
