% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classiFunc.R
\name{classiFunc}
\alias{classiFunc}
\title{The classiFunc package}
\description{
This package implements methods for functional data
classification. The main functions of this package are
\code{\link{classiKnn}}, a k nearest neighbor estimator for functional data,
and \code{\link{classiKernel}}, a kernel estimator
for functional data. The package uses efficiently implemented semimetrics to
create the distance matrix of the observations in the function
\code{\link{computeDistMat}}.
Currently supported distance measures are all \code{methods} implemented in
\code{\link[proxy]{dist}}
and all semimetrics suggested in
Fuchs etal. 2015, Nearest neighbor ensembles for functional data with
interpretable feature selection,
(\url{http://www.sciencedirect.com/science/article/pii/S0169743915001100})
Additionally, all (semi-)metrics can be used on an arbitrary order of
derivation.
This is a new package, please report all bugs and issues at
\url{https://github.com/maierhofert/classiFunc}.
}
