% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDistMat.R
\name{computeDistMat}
\alias{computeDistMat}
\title{Compute a distance matrix for functional observations}
\usage{
computeDistMat(x, y = NULL, method = "Euclidean", dmin = 1L,
  dmax = ncol(x), dmin1 = 1L, dmax1 = ncol(x), dmin2 = 1L,
  dmax2 = ncol(x), t1 = 1L, t2 = ncol(x), .poi = 1:ncol(x),
  custom.metric = function(x, y, ...) {     return(sqrt(sum((x - y)^2))) },
  ...)
}
\arguments{
\item{x}{matrix containing observations as rows}

\item{y}{see \code{x}. The default \code{NULL} uses \code{y = x}.}

\item{method}{character string specifying the (semi-)metric to be used.
on the domain of \code{x}, used in different (semi-)metrics.}

\item{dmin, dmax, dmin1, dmax1, dmin2, dmax2}{indizes used to define subspaces for
\code{method \%in\% c("shortEuclidean", "relAreas")}}

\item{t1, t2}{indizes of the points for which to compare the jump heights in
\code{method = "jump"}.}

\item{.poi}{integer vector giving the indizes of the points of interest for
\code{method = "points"}.}

\item{...}{additional parameters to the (semi-)metrics.}
}
\description{
This mainly internal function offers a unified framework to acces the
\code{\link[proxy]{dist}} function from the \code{proxy} package and additional
(semi-)metrics. For implemented methods see \code{\link{classiKnn}}.
}
