% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Growth}
\alias{Growth}
\title{Berkeley Growth Study Data}
\format{A list made up of \describe{
\item{ID}{Factor of length 93 containing the subject IDs}
\item{sex}{Factor encoding the sex of children with values in \code{c(`male`, `female`)}}
\item{age}{Numeric vector of length 31 encoding the age at the measurements}
\item{height}{A 93 x 31 matrix giving the height in cm of 93 children at 31 ages}
}}
\description{
A data frame containing the heights of 39 boys and 54 girls from age 1 to 18,
the ages at which they were collected.
}
\details{
Data and description was reformatted from the \code{\link[fda]{fda}} package.

The ages are not equally spaced, see \code{Growth$age}.
}
\references{
Ramsay, James O., and Silverman, Bernard W. (2006),
Functional Data Analysis, 2nd ed., Springer, New York.

Ramsay, James O., and Silverman, Bernard W. (2002),
Applied Functional Data Analysis, Springer, New York, ch. 6.

Tuddenham, R. D., and Snyder, M. M. (1954)
"Physical growth of California boys and girls from birth to age 18",
 University of California Publications in Child Development, 1, 183-364.
}
